+++ R/load.R
@@ -20,32 +20,6 @@
   # Scheme defaults are fairly complex...
 
   check_assumptions()
-
-  # Default options; beware of defining default options that may have different
-  # values during package install, which is when this list is contructed, and
-  # function runtime
-
-  .default.opts <- list(
-    fansi.tabs.as.spaces=FALSE,
-    fansi.tab.stops=8L,
-    fansi.warn=TRUE,
-    fansi.ctrl="all",
-    fansi.term.cap=c(
-      if(isTRUE(Sys.getenv('COLORTERM') %in% c('truecolor', '24bit')))
-      'truecolor',
-      'bright', '256'
-    ),
-    # This is not a particularly good default setting as it may exceed or fail
-    # to cover the interline distance when two lines have background colors.  To
-    # ensure lines are exactly touching use inline-block, although that has it's
-    # own issues.  Otherwise specify your own values.
-
-    fansi.css="PRE.fansi SPAN {padding-top: .25em; padding-bottom: .25em};"
-  )
-  # Scheme defaults are fairly complex...
-
-  existing.opts <- options()
-  options(.default.opts[setdiff(names(.default.opts), names(existing.opts))])
   R.ver.gte.3.2.2 <<- getRversion() >= "3.2.2"
 }
 .onAttach <- function(libname, pkgname) {